requirements.txt
black==25.1.0
bleach==6.2.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.3
databases==0.9.0
faker==37.3.0
greenlet==3.2.3
idna==3.10
iniconfig==2.1.0
isort==6.0.1
kaggle==1.7.4.5
kagglehub==0.3.12
mypy-extensions==1.1.0
mysql-connector==2.2.9
numpy==2.3.0
packaging==25.0
pandas==2.3.0
pathspec==0.12.1
pdfminer-six==20250327
pdfplumber==0.11.6
pillow==11.2.1
platformdirs==4.3.8
pluggy==1.6.0
protobuf==6.31.1
pycparser==2.22
pygments==2.19.1
pymysql==1.1.1
pypdf2==3.0.1
pypdfium2==4.30.1
pyqt6==6.9.1
pyqt6-qt6==6.9.1
pyqt6-sip==13.10.2
pytest==8.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-slugify==8.0.4
pytz==2025.2
pyyaml==6.0.2
requests==2.32.3
ruff==0.11.13
setuptools==80.9.0
six==1.17.0
sqlalchemy==2.0.41
text-unidecode==1.3
tqdm==4.67.1
typing-extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
webencodings==0.5.1
